refactor(cla): move the CLA service base URL into its own helper
Two unrelated services need claServiceBaseUrl() now that the Org Lens
EasyCLA module calls the CLA service. Importing the Me-lens cla.service
to reach one pure string function would give the Org Lens module a
dependency on Auth0 and email-verification wiring it does not use.

Move it to helpers/cla-service-url.helper.ts and re-export it from
cla.service so existing importers are unchanged.
